+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.1.0
+
+- Added Embedded Messaging API (`embeddedMessagingGetMessages`, `embeddedMessagingSetAsRead`, `embeddedMessagingReportClickMetric`, `embeddedMessagingDeleteMessage`).
+- Android: updated Optimove Android SDK to `7.11.1`.
+- iOS: update iOS SDK to `6.4.1`
+
 ## 3.0.1
 
 - iOS: Fixed duplicate interface definition error for `OptimoveInitializer` when building with React Native 0.79.6+
@@ -57,7 +63,7 @@ Fixed the parsing of the push data passed from Android to React Native
 
 - Support iOS projects with `use_frameworks!` enabled
 - [Breaking] For push opens don't set launch intent flags. Instead rely on the default React Native `singleTask` launch mode.
-- [Breaking] Changed the ```DeepLink``` object structure received from the DeepLinkHandler when calling ```Optimove.setDeepLinkHandler``` to:
+- [Breaking] Changed the `DeepLink` object structure received from the DeepLinkHandler when calling `Optimove.setDeepLinkHandler` to:
 
 ```typescript
 export interface DeepLinkContent {
@@ -74,7 +80,7 @@ interface MatchedResolution {
   resolution: DeepLinkResolution.LINK_MATCHED;
   url: string;
   content: DeepLinkContent;
-  linkData: Record<string,any>;
+  linkData: Record<string, any>;
 }
 
 export type DeepLink = NonMatchedResolution | MatchedResolution;
@@ -82,7 +88,7 @@ export type DeepLink = NonMatchedResolution | MatchedResolution;
 
 ## 1.1.1
 
-Fixed the ```Date``` types of ```InAppInboxItem``` to be ```string```
+Fixed the `Date` types of `InAppInboxItem` to be `string`
 
 ## 1.1.0